+++ demo/js/lucca-spe.js
@@ -77,4 +77,13 @@
 			return _.contains(user.firstName.toLowerCase() + user.lastName.toLowerCase(), 't') ;
 		}
 	}]);
+
+	angular.module("demoApp")
+	.controller("translationsCtrl", ["$scope", function($scope){
+		$scope.myTrads = {en: "some stuff", fr: "des bidules"};
+		$scope.count = 0
+		$scope.changed = function(){
+			$scope.count++;
+		}
+	}]);
 })();